refactor to address comment
