+++ tests/kvbm_integration/test_determinism_agg.py
@@ -22,22 +22,15 @@ fewer iterations, run the following command (expected to finish in ~58s + ~152s)
 
 import logging
 import os
-import signal
-import subprocess
-import sys
-import threading
-import time
-from datetime import datetime
 from pathlib import Path
-from typing import Any, Dict, List, Optional, TextIO
+from typing import Optional
 
 import pytest
 import requests
 
-from tests.utils.port_utils import allocate_port, deallocate_port
 from tests.utils.test_output import resolve_test_output_path
 
-from .common import DeterminismTester, ServerType
+from .common import DeterminismTester, LLMServerManager, ServerType
 from .common import TestDeterminism as BaseTestDeterminism
 from .common import check_module_available
 
@@ -65,291 +58,6 @@ pytestmark = [
     pytest.mark.gpu_1,
     pytest.mark.nightly,
 ]
-
-
-class LLMServerManager:
-    """Manages LLM server lifecycle for determinism testing."""
-
-    def __init__(
-        self,
-        base_url: Optional[str] = None,
-        port: Optional[int] = None,
-        cpu_cache_blocks: Optional[int] = None,
-        gpu_cache_blocks: Optional[int] = None,
-        log_dir: Optional[Path] = None,
-        server_type: Optional[str] = ServerType.vllm,
-    ):
-        self.server_type = server_type
-        # Use provided port, env var, or allocate a dynamic port to avoid conflicts
-        if port is not None:
-            self.port = port
-            self.port_allocated = False  # Port provided by caller, don't deallocate
-        elif os.environ.get("KVBM_SERVER_PORT"):
-            self.port = int(os.environ["KVBM_SERVER_PORT"])
-            self.port_allocated = False  # Port from env var, don't deallocate
-        else:
-            self.port = allocate_port(start_port=8000)
-            self.port_allocated = True  # Port allocated by us, must deallocate
-        self.base_url = base_url or f"http://localhost:{self.port}"
-        self.metrics_port = allocate_port(start_port=6880)
-        self.metrics_port_allocated = True
-        self.process: Optional[subprocess.Popen] = None
-        self.cpu_cache_blocks = cpu_cache_blocks
-        self.gpu_cache_blocks = gpu_cache_blocks
-
-        # Prepare logging
-        self.log_dir = log_dir or Path(".")
-        self.log_dir.mkdir(parents=True, exist_ok=True)
-        timestamp = datetime.now().strftime("%Y%m%d_%H%M%S")
-        config_str = (
-            f"cpu{cpu_cache_blocks or 'default'}_gpu{gpu_cache_blocks or 'default'}"
-        )
-        self.server_log_file = (
-            self.log_dir / f"{self.server_type}_server_{config_str}_{timestamp}.log"
-        )
-        self.server_stdout_file: Optional[TextIO] = None
-        self._tee_threads: List[threading.Thread] = []
-
-        # Environment for the process
-        self.env = os.environ.copy()
-        self.env.update(
-            {
-                "RUST_BACKTRACE": "1",
-                # DynamoConnector connection settings
-                "NATS_SERVER": "nats://localhost:4222",
-                "ETCD_ENDPOINTS": "http://localhost:2379",
-                # Enable KVBM metrics for monitoring offload/onboard
-                "DYN_KVBM_METRICS": "true",
-                "DYN_KVBM_METRICS_PORT": str(self.metrics_port),
-                # Enable vLLM batch invariant for deterministic batching
-                "VLLM_BATCH_INVARIANT": "1",
-            }
-        )
-
-        # CPU cache blocks override via env
-        if cpu_cache_blocks is not None:
-            self.env["DYN_KVBM_CPU_CACHE_OVERRIDE_NUM_BLOCKS"] = str(cpu_cache_blocks)
-
-        if self.server_type == ServerType.vllm:
-            self._set_up_vllm_config(gpu_cache_blocks)
-        elif self.server_type == ServerType.trtllm:
-            self._set_up_trtllm_config(gpu_cache_blocks)
-        else:
-            raise ValueError(
-                f"{self.server_type} is not supported yet in the KVBM test suite"
-            )
-
-    def _set_up_vllm_config(self, gpu_cache_blocks):
-        self.env["VLLM_SERVER_DEV_MODE"] = "1"
-
-        # Construct serve command
-        self.server_cmd = [
-            "vllm",
-            "serve",
-            "--block-size",
-            "16",
-            "--port",
-            str(self.port),
-            "--kv-transfer-config",
-            '{"kv_connector":"DynamoConnector","kv_role":"kv_both", "kv_connector_module_path": "kvbm.vllm_integration.connector"}',
-            os.environ.get("KVBM_MODEL_ID", "deepseek-ai/DeepSeek-R1-Distill-Llama-8B"),
-            "--attention-config.backend",
-            "FLASH_ATTN",
-            "--max-model-len",
-            "8000",  # required to fit on L4 GPU when using 8b model
-        ]
-
-        # GPU blocks override
-        if gpu_cache_blocks is not None:
-            self.server_cmd.extend(["--num-gpu-blocks-override", str(gpu_cache_blocks)])
-
-    def _set_up_trtllm_config(self, gpu_cache_blocks):
-        config_path = os.environ.get(
-            "KVBM_TRTLLM_LLMAPI_CONFIG_PATH", "/tmp/kvbm_llm_api_config.yaml"
-        )
-        llm_api_config: Dict[str, Any] = {}
-        llm_api_config[
-            "cuda_graph_config"
-        ] = None  # explicitly disable CUDA graph since Connector API doesn't support CUDA graph yet in TRTLLM
-        llm_api_config["kv_cache_config"] = {
-            "enable_partial_reuse": False,
-            "free_gpu_memory_fraction": 0.10,  # Set a small GPU fraction so that we can evict/reset the on-device kv cache faster
-        }
-        llm_api_config["kv_connector_config"] = {
-            "connector_module": "kvbm.trtllm_integration.connector",
-            "connector_scheduler_class": "DynamoKVBMConnectorLeader",
-            "connector_worker_class": "DynamoKVBMConnectorWorker",
-        }
-
-        # GPU blocks override
-        if gpu_cache_blocks is not None:
-            del llm_api_config["kv_cache_config"]["free_gpu_memory_fraction"]
-            llm_api_config["kv_cache_config"]["max_tokens"] = (
-                int(gpu_cache_blocks) * 32
-            )  # TRTLLM defaults 32 tokens per block
-
-        # Construct serve command
-        self.server_cmd = [
-            "trtllm-serve",
-            os.environ.get("KVBM_MODEL_ID", "deepseek-ai/DeepSeek-R1-Distill-Llama-8B"),
-            "--host",
-            "localhost",
-            "--port",
-            str(self.port),
-            "--backend",
-            "pytorch",
-            "--extra_llm_api_options",
-            config_path,
-        ]
-
-        import yaml
-
-        with open(config_path, "w") as f:
-            yaml.dump(llm_api_config, f, default_flow_style=False, sort_keys=False)
-
-    def _tee_output(self, pipe: Any, log_file: TextIO, prefix: str) -> None:
-        """Read from pipe and write to both log file and stdout (tee)."""
-        try:
-            for line in iter(pipe.readline, ""):
-                if not line:
-                    break
-                # Write to log file
-                log_file.write(line)
-                log_file.flush()
-                # Write to stdout with prefix
-                sys.stdout.write(f"[{prefix}] {line}")
-                sys.stdout.flush()
-        except (ValueError, OSError):
-            pass  # Pipe closed
-        finally:
-            pipe.close()
-
-    def start_server(self, timeout: int = 300) -> bool:
-        """Start LLM server and wait for readiness."""
-        if self.is_server_running():
-            self.stop_server()
-            time.sleep(2)
-
-        # Open log file (combined stdout+stderr)
-        self.server_stdout_file = open(self.server_log_file.with_suffix(".log"), "w")
-
-        # Write header
-        header = f"=== {self.server_type} Server Started at {datetime.now()} ===\nCommand: {' '.join(self.server_cmd)}\n"
-        self.server_stdout_file.write(header)
-        self.server_stdout_file.flush()
-        print(f"[{self.server_type}] {header}", end="")
-
-        # Launch with pipe, redirect stderr to stdout
-        self.process = subprocess.Popen(
-            self.server_cmd,
-            stdout=subprocess.PIPE,
-            stderr=subprocess.STDOUT,  # Redirect stderr to stdout
-            env=self.env,
-            preexec_fn=os.setsid,
-            text=True,
-            bufsize=1,  # Line buffered
-        )
-
-        # Start tee thread for combined output
-        self._tee_threads = [
-            threading.Thread(
-                target=self._tee_output,
-                args=(self.process.stdout, self.server_stdout_file, self.server_type),
-                daemon=True,
-            ),
-        ]
-        for t in self._tee_threads:
-            t.start()
-
-        # Wait for health
-        start_time = time.time()
-        while time.time() - start_time < timeout:
-            if self.is_server_running():
-                # Verify metrics endpoint is reachable (fail fast on wrong port)
-                try:
-                    requests.get(
-                        f"http://localhost:{self.metrics_port}/metrics", timeout=5
-                    )
-                    return True
-                except requests.exceptions.RequestException:
-                    print(
-                        f"Warning: server healthy but metrics port {self.metrics_port} not reachable yet"
-                    )
-            if self.process.poll() is not None:
-                # Process exited, wait for tee thread to finish
-                for t in self._tee_threads:
-                    t.join(timeout=2)
-                self._close_log_files()
-                return False
-            time.sleep(5)
-
-        # Timeout
-        self.stop_server()
-        return False
-
-    def stop_server(self):
-        """Stop LLM server and close logs."""
-        if self.process:
-            try:
-                os.killpg(os.getpgid(self.process.pid), signal.SIGTERM)
-                try:
-                    self.process.wait(timeout=30)
-                except subprocess.TimeoutExpired:
-                    os.killpg(os.getpgid(self.process.pid), signal.SIGKILL)
-                    self.process.wait()
-            except (ProcessLookupError, OSError):
-                pass
-            finally:
-                self.process = None
-        # Wait for tee threads to finish
-        for t in self._tee_threads:
-            t.join(timeout=2)
-        self._tee_threads = []
-        self._close_log_files()
-
-        # Deallocate ports if we allocated them
-        if self.port_allocated:
-            deallocate_port(self.port)
-            self.port_allocated = False
-        if self.metrics_port_allocated:
-            deallocate_port(self.metrics_port)
-            self.metrics_port_allocated = False
-
-    def _close_log_files(self):
-        if self.server_stdout_file:
-            self.server_stdout_file.write(
-                f"\n=== Server Stopped at {datetime.now()} ===\n"
-            )
-            self.server_stdout_file.close()
-            self.server_stdout_file = None
-
-    def is_server_running(self) -> bool:
-        try:
-            # First check basic health
-            response = requests.get(f"{self.base_url}/health", timeout=5)
-            if response.status_code != 200:
-                return False
-
-            # Then check if the model endpoint is ready with a simple test request
-            test_payload = {
-                "model": os.environ.get(
-                    "KVBM_MODEL_ID", "deepseek-ai/DeepSeek-R1-Distill-Llama-8B"
-                ),
-                "messages": [{"role": "user", "content": "test"}],
-                "max_completion_tokens": 1,
-                "temperature": 0,
-            }
-
-            response = requests.post(
-                f"{self.base_url}/v1/chat/completions",
-                headers={"Content-Type": "application/json"},
-                json=test_payload,
-                timeout=10,
-            )
-            return response.status_code == 200
-
-        except requests.exceptions.RequestException:
-            return False
 
 
 class AggDeterminismTester(DeterminismTester):
